backend/requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# Database
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
pgvector>=0.2.5
alembic>=1.13.1

# LLM / Embeddings
openai>=1.30.0

# Document parsing
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
python-docx>=1.1.0

# Utilities
python-multipart>=0.0.9
httpx>=0.27.0
tenacity>=8.3.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
